Graph refactor
